# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing community waste collection — households, pickups, payments, reports.",
        "title": "Community Waste Collection API",
        "contact": {
            "name": "Muhammad Faisal Affan"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/households": {
            "get": {
                "description": "Get paginated list of households",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Households"
                ],
                "summary": "List households",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    }
                }
            },
            "post": {
                "description": "Register a new household",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Households"
                ],
                "summary": "Create household",
                "parameters": [
                    {
                        "description": "Household data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_internal_domain.CreateHouseholdRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_internal_domain.Household"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
        "/households/{id}": {
            "get": {
                "description": "Get household by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Households"
                ],
                "summary": "Get household",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Household UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_internal_domain.Household"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete household by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Households"
                ],
                "summary": "Delete household",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Household UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_faisalaffan_community-waste-collection-api_internal_domain.CreateHouseholdRequest": {
            "type": "object",
            "required": [
                "address",
                "owner_name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "owner_name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_faisalaffan_community-waste-collection-api_internal_domain.Household": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "owner_name": {
                    "type": "string"
                },
                "payments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_internal_domain.Payment"
                    }
                },
                "pickups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_internal_domain.WastePickup"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_faisalaffan_community-waste-collection-api_internal_domain.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "household_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payment_date": {
                    "type": "string"
                },
                "proof_file_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "waste_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_faisalaffan_community-waste-collection-api_internal_domain.WastePickup": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "household_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "pickup_date": {
                    "type": "string"
                },
                "safety_check": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.apiError"
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.pagination"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_faisalaffan_community-waste-collection-api_pkg_response.ValidationDetail": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_faisalaffan_community-waste-collection-api_pkg_response.apiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.ValidationDetail"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_faisalaffan_community-waste-collection-api_pkg_response.pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        }
    }
}